refactor new patient exam to separate component
